// src/common/types.ts
export interface PlayerUpdate {
  readonly frameNumber: number;
  readonly playerIndex: number;
  readonly inputs: PlayerInputs;
  readonly nanaInputs?: PlayerInputs;
  readonly state: PlayerState;
  readonly nanaState?: PlayerState;
}

export type PlayerUpdateWithNana = Required<PlayerUpdate>;

export interface PlayerInputs {
  readonly frameNumber: number;
  readonly playerIndex: number;
  readonly isNana: boolean;
  readonly physical: {
    readonly a: boolean;
    readonly b: boolean;
    readonly x: boolean;
    readonly y: boolean;
    readonly z: boolean;
    readonly start: boolean;
    readonly dPadLeft: boolean;
    readonly dPadRight: boolean;
    readonly dPadDown: boolean;
    readonly dPadUp: boolean;
    /** range [0, 1] */
    readonly rTriggerAnalog: number;
    readonly rTriggerDigital: boolean;
    /** range [0, 1] */
    readonly lTriggerAnalog: number;
    readonly lTriggerDigital: boolean;
  };
  readonly processed: {
    readonly a: boolean;
    readonly b: boolean;
    readonly x: boolean;
    readonly y: boolean;
    readonly z: boolean;
    readonly start: boolean;
    readonly dPadLeft: boolean;
    readonly dPadRight: boolean;
    readonly dPadDown: boolean;
    readonly dPadUp: boolean;
    readonly rTriggerDigital: boolean;
    readonly lTriggerDigital: boolean;
    /** range [-1, 1] */
    readonly joystickX: number;
    /** range [-1, 1] */
    readonly joystickY: number;
    /** range [-1, 1] */
    readonly cStickX: number;
    /** range [-1, 1] */
    readonly cStickY: number;
    /** range [0, 1] */
    readonly anyTrigger: number;
  };
}

export interface PlayerState {
  readonly frameNumber: number;
  readonly playerIndex: number;
  readonly isNana: boolean;
  readonly internalCharacterId: number;
  readonly actionStateId: number;
  readonly xPosition: number;
  readonly yPosition: number;
  readonly facingDirection: number;
  readonly percent: number;
  readonly shieldSize: number;
  readonly lastHittingAttackId: number;
  readonly currentComboCount: number;
  readonly lastHitBy: number;
  readonly stocksRemaining: number;
  readonly actionStateFrameCounter: number;
  readonly hitstunRemaining: number;
  readonly isGrounded: boolean;
  readonly lastGroundId: number;
  readonly jumpsRemaining: number;
  /**
   * L-cancel status is only populated on the first landing frame, not all of
   * them.
   * */
  readonly lCancelStatus: "successful" | "missed" | undefined;
  readonly hurtboxCollisionState: "vulnerable" | "invulnerable" | "intangible";
  readonly selfInducedAirXSpeed: number;
  readonly selfInducedAirYSpeed: number;
  readonly attackBasedXSpeed: number;
  readonly attackBasedYSpeed: number;
  readonly selfInducedGroundXSpeed: number;
  readonly hitlagRemaining: number;
  readonly isReflectActive: boolean;
  readonly isFastfalling: boolean;
  readonly isShieldActive: boolean;
  readonly isInHitstun: boolean;
  readonly isHittingShield: boolean;
  readonly isPowershieldActive: boolean;
  readonly isDead: boolean;
  readonly isOffscreen: boolean;
}

export interface ItemUpdate {
  readonly frameNumber: number;
  readonly typeId: number;
  readonly state: number;
  readonly facingDirection: number;
  readonly xVelocity: number;
  readonly yVelocity: number;
  readonly xPosition: number;
  readonly yPosition: number;
  readonly damageTaken: number;
  readonly expirationTimer: number;
  readonly spawnId: number;
  readonly samusMissileType: number;
  readonly peachTurnipFace: number;
  /** Mewtwo/Samus */
  readonly isChargeShotLaunched: boolean;
  /** Mewtwo/Samus */
  readonly chargeShotChargeLevel: number;
  readonly owner: number;
}